Refactoring

+++ smartvision_system/core/protocol.cc
@@ -12,7 +12,7 @@ void parseSetZoomCmd(const std::vector<uint8_t> &command, float *zoomFactor, flo
     *zoomFactor = desiredZoom;
 }
 
-void parseSetGimbalCmd(const std::vector<uint8_t> &command, uint8_t *panAngle, uint8_t *tiltAngle, uint8_t *reset) {
+void parseSetGimbalCmd(const std::vector<uint8_t> &command, int *panAngle, int *tiltAngle, uint8_t *reset) {
     if (command[1] == 43) {
         *panAngle += 5;
     } else if (command[1] == 45) {
